+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ### Added
 * Added option to mask nodata regions when using imread and gdal backend.
+* Added area property to Polygon and MultiPolygon.
 
 
 ## Version 0.8.1 - Released 2022-03-04