+++ CHANGELOG.md
@@ -4,7 +4,21 @@ This changelog follows the specifications detailed in: [Keep a Changelog](https:
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html), although we have not yet reached a `1.0.0` release.
 
 
-## Version 0.8.3 - Unreleased
+## Version 0.8.4 - Unreleased
+
+### Fixed
+* Bug in `Detections.draw_on(..., color='classes')` when a class is unknown
+* Added workaround to cv2 bug to `Polygon.draw_on` when polygon coordinates are
+  too large. Currently clips the polygon and emits a warning, this prevents
+  crashes, but results may not be correct.
+
+
+### Changed
+* `Boxes.from_slice` no longer handles wraparound slices as it is not well
+  defined and previous behavior was buggy.
+
+
+## Version 0.8.3 - Released 2022-03-28
 
 ### Added
 * facecolor and edgecolor in `Polygon.draw_on`